Add input protection for process classes Fix bugs in geometry classes

+++ include/G2PBPMTrans.hh
@@ -21,42 +21,42 @@
 
 namespace Orbit1
 {
-float target_x(float *x, int m);
-float target_y(float *x, int m);
-float target_theta(float *x, int m);
-float target_phi(float *x, int m);
+double target_x(double *x, int m);
+double target_y(double *x, int m);
+double target_theta(double *x, int m);
+double target_phi(double *x, int m);
 }
 
 namespace Orbit4
 {
-float target_x(float *x, int m);
-float target_y(float *x, int m);
-float target_theta(float *x, int m);
-float target_phi(float *x, int m);
+double target_x(double *x, int m);
+double target_y(double *x, int m);
+double target_theta(double *x, int m);
+double target_phi(double *x, int m);
 }
 
 namespace Orbit5
 {
-float target_x(float *x, int m);
-float target_y(float *x, int m);
-float target_theta(float *x, int m);
-float target_phi(float *x, int m);
+double target_x(double *x, int m);
+double target_y(double *x, int m);
+double target_theta(double *x, int m);
+double target_phi(double *x, int m);
 }
 
 namespace Orbit7
 {
-float target_x(float *x, int m);
-float target_y(float *x, int m);
-float target_theta(float *x, int m);
-float target_phi(float *x, int m);
+double target_x(double *x, int m);
+double target_y(double *x, int m);
+double target_theta(double *x, int m);
+double target_phi(double *x, int m);
 }
 
 namespace Orbit9
 {
-float target_x(float *x, int m);
-float target_y(float *x, int m);
-float target_theta(float *x, int m);
-float target_phi(float *x, int m);
+double target_x(double *x, int m);
+double target_y(double *x, int m);
+double target_theta(double *x, int m);
+double target_phi(double *x, int m);
 }
 
 #endif
